// workspaces/server/src/types.ts
export interface RedisConfig {
  port?: number
  host?: string
  family?: number
  password?: string
  db?: number
}

export interface WorkerConfig {
  prefix: string
  version: string
  db: DBConfig
  api: { [api: string]: string }
  storageService: string
  shapesContainer: string
  shapesRegion: string
  newRealtime: boolean
  redis: RedisConfig
}

export interface EcsGatewayConfig {
  cluster: string
  region: string
  servicePrefix: string
  serviceSuffix: string
  replicas: number
}

export interface KubernetesGatewayConfig {
  deploymentSuffix: string
  deploymentPrefix: string
  namespace: string
  containerName: string
  replicas: number
}

export interface DBConfig extends SqlConfig {
  server: string
  user: string
  password: string
}